chore: remove individual test files. bring back public-methods.test.ts
